refactored autocomplete file structure
