+++ express/blocks/search-marquee/search-marquee.js
@@ -174,7 +174,7 @@ function initSearchFunction(block) {
     }
   };
 
-  import('./use-input-autocomplete.js').then(({ default: useInputAutocomplete }) => {
+  import('../../scripts/autocomplete-api-v3.js').then(({ default: useInputAutocomplete }) => {
     const { inputHandler } = useInputAutocomplete(
       suggestionsListUIUpdateCB, { throttleDelay: 300, debounceDelay: 500, limit: 7 },
     );
